requirements.txt
jsonschema
simplejson
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
requests==2.18.1
six==1.12.0
urllib3==1.21.1
Werkzeug==0.12.2
pymongo==3.4.0
gensim==2.3.0
cloudpickle
numpy
jieba
docopt
scikit-learn==0.18.1
django==2.0
Markdown==2.6.10
django-cors-headers==2.1.0
djangorestframework==3.7.3
pytest
